requirements.txt
altgraph==0.17.2
autopep8==1.6.0
colorama==0.4.4
future==0.18.2
global-hotkeys==0.0.4
numpy==1.21.2
opencv-python==4.5.5.62
pefile==2021.9.3
Pillow==9.0.1
pycodestyle==2.8.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pywin32==302
pywin32-ctypes==0.2.0
toml==0.10.2
tqdm==4.63.0
